third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-shown-node-count-preserved-when-sorting-expected.txt

Tests Summary view of detailed heap snapshots. Shown node count must be preserved after sorting.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testExpansionPreservedWhenSorting

Profiler was disabled.
